requirements.txt
absl-py==2.1.0
accelerate==0.12.0
annotated-types==0.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.2
contextlib2==21.6.0
contourpy==1.3.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.7.0
emoji==2.14.0
eval_type_backport==0.2.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
ftfy==6.1.1
gitdb==4.0.11
GitPython==3.1.43
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.8
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.10.0
ml_collections==0.1.1
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
open-clip-torch==2.23.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.29.1
psutil==6.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.1
pyre-extensions==0.0.23
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scipy==1.8.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
six==1.17.0
smmap==5.0.1
stanza==1.10.1
sympy==1.13.3
timm==0.9.7
tokenizers==0.15.2
tomli==2.2.1
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.67.1
transformers==4.35.2
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.1
wcwidth==0.2.13
xformers==0.0.17
